LRIS-31876 port of LRIS-31164 format source and fix compile errors

+++ modules/dxp/apps/osb/osb-customer/osb-portlet/docroot/WEB-INF/service/com/liferay/osb/model/ProductEntryConstants.java
@@ -357,7 +357,7 @@ public class ProductEntryConstants {
 	}
 
 	public static boolean isDigitalEnterpriseVersion7_1(
-		int digitalEnterpriseVersion) {
+		long digitalEnterpriseVersion) {
 
 		if (digitalEnterpriseVersion == DIGITAL_ENTERPRISE_VERSION_7_1_10) {
 			return true;
